Better unit conversion and test cases
